requirements.txt
# Timmy-IA Requirements
# ✅ CORRIGIDO: Dependências verificadas e organizadas

# Core LLM
openai>=1.30.0

# Environment & Config  
python-dotenv>=1.0.0

# Web Framework (Streamlit)
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Future integrations
# twilio  # Para WhatsApp (futuro)
# redis   # Para cache (futuro)
# psycopg2-binary  # Para PostgreSQL (futuro)